feature/pretty-snapshot > Rectified collisions when merging features like pretty-format & custom stringifyFn
This commit fixes the collisions between the different merges.  The module
now always prints to the snapshot files in pretty-format while being able to
handle a custom serialize function which may use JSON.stringify. The module
now is now consistent from what it was in this branch.

Maintained most of the test cases but updated all snapshots.